# contract/openapi.yaml
openapi: 3.1.0
info:
  title: SaaS Starter API
  version: 1.0.0
  description: |
    Shared API contract for the SaaS Starter Kit.
    All backend implementations (Go, Python, Node.js, etc.) must conform to this spec.
servers:
  - url: http://localhost:8080
    description: Default development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, role, verified, two_fa_enabled]
      properties:
        id:
          type: string
          description: Public UUID (never the internal serial ID)
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        language:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        verified:
          type: boolean
        two_fa_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────
...
    AdminCreateUserRequest:
      type: object
      required: [first_name, last_name, email, password]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 255
        role:
          type: string
        language:
          type: string
        verified:
          type: boolean
    AdminUpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 255
        role:
          type: string
        language:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        verified:
          type: boolean
        two_fa_enabled:
          type: boolean
    AdminUpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, user]
...
    AdminAction:
      type: object
      properties:
        id:
          type: integer
        admin_id:
          type: integer
        admin_email:
          type: string
        action:
          type: string
        target_user_id:
          type: integer
        target_email:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
    PaginatedUsers:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin - Users]
      summary: List all users (paginated)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: sort_field
          in: query
          schema:
            type: string
            default: id
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
    post:
      tags: [Admin - Users]
      summary: Create a new user
      operationId: adminCreateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{id}:
    get:
      tags: [Admin - Users]
      summary: Get user by ID
      operationId: adminGetUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Admin - Users]
      summary: Update user
      operationId: adminUpdateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [Admin - Users]
      summary: Delete user
      operationId: adminDeleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{id}/role:
    put:
      tags: [Admin - Users]
      summary: Update user role
      operationId: adminUpdateUserRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/admin/actions:
    get:
      tags: [Admin - Users]
      summary: Admin audit log
      operationId: getAdminActions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Admin actions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminAction"
                  total:
                    type: integer

  # ── Admin - Analytics ─────────────────────────────────
...